get pad stuff working
This was a pain, but I think it's all working now. Everything gets
drawn on a pad, and when the window gets resized, or a drawing is
loaded that was made on a bigger screen, the pad gets redisplayed,
making an effort to include as much of the drawing as possible.